+++ src/Blackjack/BlackjackPlayer.cpp
@@ -100,7 +100,7 @@ void BlackjackPlayer::UpdateCardPosition(CardHand* targetHand)
         Card* cardIter = (*it);
 
         FPosition cardPos;
-        cardPos.X = 200 + (100 * cardIndex);
+        cardPos.X = 200 + (100 * (float)cardIndex);
         cardPos.Y = 400;
 
         cardIter->cardFront->SetPos(cardPos);
